requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
